fill label holes

+++ src/vollseg/utils.py
@@ -4381,7 +4381,7 @@ def CellPoseWater(membrane_image, sized_smart_seeds, mask, decay_rate = 1):
     #    for z in range(z_dim):
     #        membrane_image[z] *= decay_map[z]
     
-    labeled_image = np.zeros(membrane_image.shape)
+    labeled_image = np.zeros(membrane_image.shape, dtype=np.uint16)
     for i in range(membrane_image.shape[0]):
        labeled_image[i] = label(binary_image[i])
